update Item unit_proce to float

+++ db/migrate/20220515185912_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[7.0]
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.integer :unit_price
+      t.float :unit_price
       t.references :merchant, null: false, foreign_key: true
 
       t.timestamps
